Added roles.

+++ IStudyKindergardens/Controllers/HomeController.cs
@@ -13,68 +13,57 @@ namespace IStudyKindergardens.Controllers
 {
     public class HomeController : Controller
     {
+        [HttpGet]
         public ActionResult Index()
         {
             return View();
         }
 
-        public ActionResult About()
-        {
-            ViewBag.Message = "Your application description page.";
-
-            return View();
-        }
-
-        public ActionResult Contact()
-        {
-            ViewBag.Message = "Your contact page.";
-
-            return View();
-        }
-
-
         [HttpPost]
         public JsonResult UploadPicture()
         {
-            foreach (string file in Request.Files)
+            if (User.Identity.IsAuthenticated)
             {
-                var upload = Request.Files[file];
-                if (upload != null)
+                foreach (string file in Request.Files)
                 {
-                    string expansion = upload.FileName.Split(new char[] { '.' }).Last();
-                    string fileName = Guid.NewGuid().ToString() + '.' + expansion;
-                    string path = Server.MapPath("~/Images/Uploaded/Temp/" + fileName);
-                    upload.SaveAs(path);
-                    try
+                    var upload = Request.Files[file];
+                    if (upload != null)
                     {
-                        using (MemoryStream memory = new MemoryStream())
+                        string expansion = upload.FileName.Split(new char[] { '.' }).Last();
+                        string fileName = Guid.NewGuid().ToString() + '.' + expansion;
+                        string path = Server.MapPath("~/Images/Uploaded/Temp/" + fileName);
+                        upload.SaveAs(path);
+                        try
                         {
-                            using (Image img = Image.FromFile(path))
+                            using (MemoryStream memory = new MemoryStream())
                             {
-                                PropertyItem item = img.GetPropertyItem(274);
-                                if (item.Value[0] == 3)
+                                using (Image img = Image.FromFile(path))
                                 {
-                                    img.RotateFlip(RotateFlipType.Rotate180FlipXY);
+                                    PropertyItem item = img.GetPropertyItem(274);
+                                    if (item.Value[0] == 3)
+                                    {
+                                        img.RotateFlip(RotateFlipType.Rotate180FlipXY);
+                                    }
+                                    else if (item.Value[0] == 6)
+                                    {
+                                        img.RotateFlip(RotateFlipType.Rotate270FlipXY);
+                                    }
+                                    else if (item.Value[0] == 8)
+                                    {
+                                        img.RotateFlip(RotateFlipType.Rotate90FlipXY);
+                                    }
+                                    img.Save(memory, ImageFormat.Jpeg);
                                 }
-                                else if (item.Value[0] == 6)
+                                using (FileStream fs = new FileStream(path, FileMode.Create, FileAccess.ReadWrite))
                                 {
-                                    img.RotateFlip(RotateFlipType.Rotate270FlipXY);
+                                    byte[] bytes = memory.ToArray();
+                                    fs.Write(bytes, 0, bytes.Length);
                                 }
-                                else if (item.Value[0] == 8)
-                                {
-                                    img.RotateFlip(RotateFlipType.Rotate90FlipXY);
-                                }
-                                img.Save(memory, ImageFormat.Jpeg);
-                            }
-                            using (FileStream fs = new FileStream(path, FileMode.Create, FileAccess.ReadWrite))
-                            {
-                                byte[] bytes = memory.ToArray();
-                                fs.Write(bytes, 0, bytes.Length);
                             }
                         }
+                        catch (Exception) { }
+                        return Json(fileName);
                     }
-                    catch (Exception) { }
-                    return Json(fileName);
                 }
             }
             return Json(false);
@@ -83,11 +72,14 @@ namespace IStudyKindergardens.Controllers
         [HttpPost]
         public JsonResult DeletePicture(string id)
         {
-            try
+            if (User.Identity.IsAuthenticated)
             {
-                System.IO.File.Delete(Server.MapPath("~/Images/Uploaded/Temp/") + id);
+                try
+                {
+                    System.IO.File.Delete(Server.MapPath("~/Images/Uploaded/Temp/") + id);
+                }
+                catch (Exception) { }
             }
-            catch (Exception) { }
             return Json(true);
         }
     }
